Translate tab to \t in sent massages.
When set to 'opportunistic' mode, otr suffixes messages with a bunch of spaces
and tabs, JSON does not allow tabs/line feeds/carriage returns within any
tokens, so we need to encode tabs as \t.

Leaving cariage returns and line feeds as is for now in dumb hope that noone
will ever use those(you wish).