Remove g_main_context_iteration.
Bitlbee can use both glib and libevent eventloops, but
g_main_context_iteration is glib's function so in case when bitlbee uses
libevent it would create and try to iterate a completely unrelated
eventloop.

Removing this doesn't seem to break anything as it was added at the
point where there was no mechanism to track if all servers are synced.

Should fix issue #62.