set default away status to 'idle'
Starting with 98f3893 we support multiple awat statuses, that commit
changed the default away status to "online", resulting in commands like
'/away food' working the same way as 'set status food', i.e. leaving the
person in online state, but with a message set.
Change the default away status to 'idle' as it worked previously, this
should fix #161.