Fixed crashing problems upon rejecting more than one friend request. (#172)
I think this is caused by uinf and uinf->user being a null pointer. Not really sure why that is, but this fixes the crash that would otherwise cause. I imagine it's because the user has never actually existed within the bitlbee user list and therefore doesn't need updating, but the real question is why it works for the first rejection but not the second.

I also slightly cleaned it up so uinf is defined before usage in either case, meaning there's only one definition statement.