Don't wait for GUILD_SYNC on login
For whatever reason discord stopped responding to REQUEST_SYNC
opcode(12) so we timeout on login. Temporary ignore the issue and assume
we have completed connections even though we are missing guild info.

I am no longer sure of the full list of effects this will have, we will
surely lose guild presence info.

This is a workaround for #201 and should be reverted once that is
properly fixed.

Signed-off-by: Artem Savkov <artem.savkov@gmail.com>